move delta stuff to own file
